+++ perum-tenant/app/modules/mobile_descriptor.py
@@ -54,6 +54,7 @@ class MobileCapabilities(BaseModel):
     social_realtime: StrictBool
     social_attachments: StrictBool
     support_requester: StrictBool
+    support_admin: StrictBool
     support_attachments: StrictBool
     offline_preferences: StrictBool
     offline_homework_state: StrictBool